Add tests and mechanism for running them through Travis CI
This patch adds tests for the PuzzleTools library and ensures they
are run through Travis CI to verify that they succeed. It adds
livecode as a submodule, and instructs travis to build the LiveCode
community engine and run the tests. The tests are run using the
livecode submodule's test runner via a Makefile which invokes
the livecode test Makefile with a few tweaked parameters.